pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name="freaky"
authors = [
	{ name = "Leslie Schumm", email = "lschumm@protonmail.com" }
]
description = "tools for frequency domain analysis of audio."
readme = "README.md"
license = {text = "GPL-2.0-only"}
version = "1.0.0a11"
dynamic = []
dependencies = [
	"click",
	"numpy",
	"scipy",
	"numba",
	"pillow"
	]

[project.scripts]
freaky-encode = "freaky:_encode_cli"
freaky-decode = "freaky:_decode_cli"
